Remove recursion from DLQ processor
Lambda seems to handle recursion very weirdly for scheduled events. If
the scheduled-event rate is set to 1 minute, the invoked function
sometimes (but not always?) doesn't run for another minute.

Instead, just increase the DLQ function timeout to 60 and process SQS
events via the main function until 10 seconds are left before the
timeout (which could be lower, but that's the current timeout for the
main function).